Refactor code and throw Maven/Gradle exception
